Refactor and simplify reverted matchers
