+++ packages/hardhat-chai-matchers/src/hardhatChaiMatchers.ts
@@ -1,12 +1,16 @@
 import "./types";
 import { supportBigNumber } from "./bigNumber";
 import { supportEmit } from "./emit";
-import { supportReverted } from "./reverted";
 import { supportHexEqual } from "./hexEqual";
 import { supportProperAddress } from "./properAddress";
 import { supportProperPrivateKey } from "./properPrivateKey";
 import { supportChangeEtherBalance } from "./changeEtherBalance";
 import { supportChangeEtherBalances } from "./changeEtherBalances";
+import { supportReverted } from "./reverted/reverted";
+import { supportRevertedWith } from "./reverted/revertedWith";
+import { supportRevertedWithCustomError } from "./reverted/revertedWithCustomError";
+import { supportRevertedWithPanic } from "./reverted/revertedWithPanic";
+import { supportRevertedWithoutReasonString } from "./reverted/revertedWithoutReasonString";
 
 export function hardhatChaiMatchers(
   chai: Chai.ChaiStatic,
@@ -14,10 +18,14 @@ export function hardhatChaiMatchers(
 ) {
   supportBigNumber(chai.Assertion, utils);
   supportEmit(chai.Assertion);
-  supportReverted(chai.Assertion, utils);
   supportHexEqual(chai.Assertion);
   supportProperAddress(chai.Assertion);
   supportProperPrivateKey(chai.Assertion);
   supportChangeEtherBalance(chai.Assertion);
   supportChangeEtherBalances(chai.Assertion);
+  supportReverted(chai.Assertion);
+  supportRevertedWith(chai.Assertion);
+  supportRevertedWithCustomError(chai.Assertion, utils);
+  supportRevertedWithPanic(chai.Assertion);
+  supportRevertedWithoutReasonString(chai.Assertion);
 }
