+++ packages/hardhat-chai-matchers/test/panic.ts
@@ -1,7 +1,7 @@
 import { assert } from "chai";
 import { BigNumber } from "ethers";
 
-import { PANIC_CODES, panicErrorCodeToReason } from "../src/panic";
+import { PANIC_CODES, panicErrorCodeToReason } from "../src/reverted/panic";
 
 describe("panic codes", function () {
   it("all exported panic codes should have a description", async function () {
